# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "HF3 Backend",
        "description": "A backend project made in HF3",
        "contact": {
            "email": "kristian@binau.me"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/customers/{customerId}": {
            "get": {
                "tags": [
                    "Customers"
                ],
                "summary": "Get customer",
                "description": "Returns Get customer",
                "operationId": "CustomerController.show",
                "parameters": [
                    {
                        "name": "customerId",
                        "in": "path",
                        "description": "Customer Id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                },
                "security": [
                    {
                        "api_key_security_example": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Customers"
                ],
                "summary": "Store customer",
                "description": "Stores customer and returns Get customer",
                "operationId": "CustomerController.store",
                "parameters": [
                    {
                        "name": "customerId",
                        "in": "path",
                        "description": "Customer Id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                },
                "security": [
                    {
                        "api_key_security_example": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Customers"
                ],
                "summary": "Delete customer",
                "description": "Deletes customer and returns nothing",
                "operationId": "CustomerController.destroy",
                "parameters": [
                    {
                        "name": "customerId",
                        "in": "path",
                        "description": "Customer Id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                },
                "security": [
                    {
                        "api_key_security_example": []
                    }
                ]
            },
            "patch": {
                "tags": [
                    "Customers"
                ],
                "summary": "Update customer",
                "description": "Updates customer and returns Get customer",
                "operationId": "CustomerController.update",
                "parameters": [
                    {
                        "name": "customerId",
                        "in": "path",
                        "description": "Customer Id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                },
                "security": [
                    {
                        "api_key_security_example": []
                    }
                ]
            }
        },
...
        "/api/departments/{departmentId}": {
            "get": {
                "tags": [
                    "Locations"
                ],
                "summary": "Get department",
                "description": "Returns Get department",
                "operationId": "DepartmentController.show",
                "parameters": [
                    {
                        "name": "departmentId",
                        "in": "path",
                        "description": "Department Id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                },
                "security": [
                    {
                        "api_key_security_example": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Locations"
                ],
                "summary": "Store department",
                "description": "Stores department and returns Get department",
                "operationId": "DepartmentController.store",
                "parameters": [
                    {
                        "name": "departmentId",
                        "in": "path",
                        "description": "Department Id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                },
                "security": [
                    {
                        "api_key_security_example": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Locations"
                ],
                "summary": "Delete department",
                "description": "Deletes department and returns nothing",
                "operationId": "DepartmentController.destroy",
                "parameters": [
                    {
                        "name": "departmentId",
                        "in": "path",
                        "description": "Department Id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                },
                "security": [
                    {
                        "api_key_security_example": []
                    }
                ]
            },
            "patch": {
                "tags": [
                    "Locations"
                ],
                "summary": "Update department",
                "description": "Updates department and returns Get department",
                "operationId": "DepartmentController.update",
                "parameters": [
                    {
                        "name": "departmentId",
                        "in": "path",
                        "description": "Department Id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                },
                "security": [
                    {
                        "api_key_security_example": []
                    }
                ]
            }
        },
...
        "/api/employees/{employeeId}": {
            "get": {
                "tags": [
                    "Locations"
                ],
                "summary": "Get employee",
                "description": "Returns Get employee",
                "operationId": "EmployeeController.show",
                "parameters": [
                    {
                        "name": "employeeId",
                        "in": "path",
                        "description": "Employee Id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                },
                "security": [
                    {
                        "api_key_security_example": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Locations"
                ],
                "summary": "Store employee",
                "description": "Stores employee and returns Get employee",
                "operationId": "EmployeeController.store",
                "parameters": [
                    {
                        "name": "employeeId",
                        "in": "path",
                        "description": "Employee Id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                },
                "security": [
                    {
                        "api_key_security_example": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Locations"
                ],
                "summary": "Delete employee",
                "description": "Deletes employee and returns nothing",
                "operationId": "EmployeeController.destroy",
                "parameters": [
                    {
                        "name": "employeeId",
                        "in": "path",
                        "description": "Employee Id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                },
                "security": [
                    {
                        "api_key_security_example": []
                    }
                ]
            },
            "patch": {
                "tags": [
                    "Locations"
                ],
                "summary": "Update employee",
                "description": "Updates employee and returns Get employee",
                "operationId": "EmployeeController.update",
                "parameters": [
                    {
                        "name": "employeeId",
                        "in": "path",
                        "description": "Employee Id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                },
                "security": [
                    {
                        "api_key_security_example": []
                    }
                ]
            }
        },
...
        "/api/items/{itemId}": {
            "get": {
                "tags": [
                    "Items"
                ],
                "summary": "Get item",
                "description": "Returns Get item",
                "operationId": "ItemController.show",
                "parameters": [
                    {
                        "name": "itemId",
                        "in": "path",
                        "description": "Item Id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                },
                "security": [
                    {
                        "api_key_security_example": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Items"
                ],
                "summary": "Store item",
                "description": "Stores item and returns Get item",
                "operationId": "ItemController.store",
                "parameters": [
                    {
                        "name": "itemId",
                        "in": "path",
                        "description": "Item Id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                },
                "security": [
                    {
                        "api_key_security_example": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Items"
                ],
                "summary": "Delete item",
                "description": "Deletes item and returns nothing",
                "operationId": "ItemController.destroy",
                "parameters": [
                    {
                        "name": "itemId",
                        "in": "path",
                        "description": "Item Id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                },
                "security": [
                    {
                        "api_key_security_example": []
                    }
                ]
            },
            "patch": {
                "tags": [
                    "Items"
                ],
                "summary": "Update item",
                "description": "Updates item and returns Get item",
                "operationId": "ItemController.update",
                "parameters": [
                    {
                        "name": "itemId",
                        "in": "path",
                        "description": "Item Id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                },
                "security": [
                    {
                        "api_key_security_example": []
                    }
                ]
            }
        },
...
        "/api/orders": {
            "get": {
                "tags": [
                    "Orders"
                ],
                "summary": "Get list of orders",
                "description": "Returns list of orders",
                "operationId": "OrderController.index",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Page number",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                },
                "security": [
                    {
                        "api_key_security_example": []
                    }
                ]
            }
        },
...
    },
...
    "security": [
        []
    ]
}